rename application directory
