+++ com.example.rcpapp/META-INF/MANIFEST.MF
@@ -28,8 +28,7 @@ Bundle-ClassPath: .,
  lib/guava-20.0.jar,
  lib/commons-configuration-1.10.jar,
  lib/jackson-databind-2.3.1.jar
-Export-Package: com.example.rcpapp,
- com.fasterxml.jackson.annotation,
+Export-Package: com.fasterxml.jackson.annotation,
  com.fasterxml.jackson.core,
  com.fasterxml.jackson.core.base,
  com.fasterxml.jackson.core.filter,
@@ -63,7 +62,7 @@ Export-Package: com.example.rcpapp,
  com.github.xws,
  com.google.common.annotations,
  com.google.common.base,
- com.google.common.base.internal,
+ com.google.common.base.internal;x-internal:=true,
  com.google.common.cache,
  com.google.common.collect,
  com.google.common.escape,
@@ -124,6 +123,7 @@ Export-Package: com.example.rcpapp,
  com.mchange.v2.sql,
  com.mchange.v2.sql.filter,
  com.mchange.v2.util,
+ freemarker,
  freemarker.cache,
  freemarker.core,
  freemarker.debug,
@@ -247,9 +247,13 @@ Export-Package: com.example.rcpapp,
  org.terracotta.quartz.collections,
  org.terracotta.quartz.wrappers,
  us.nineworlds.xstreamer,
+ us.nineworlds.xstreamer.forms,
  us.nineworlds.xstreamer.jobs,
  us.nineworlds.xstreamer.model,
- us.nineworlds.xstreamer.model.lookup
+ us.nineworlds.xstreamer.model.lookup,
+ us.nineworlds.xstreamer.preferences,
+ us.nineworlds.xstreamer.preferences.widgets,
+ us.nineworlds.xstreamer.ui.importWizards
 Import-Package: org.apache.commons.lang,
  org.apache.commons.lang.builder,
  org.apache.commons.lang.enum;version="2.6.0",
@@ -260,5 +264,5 @@ Import-Package: org.apache.commons.lang,
  org.apache.commons.lang.reflect,
  org.apache.commons.lang.text,
  org.apache.commons.lang.time
-Bundle-Activator: com.example.rcpapp.Activator
+Bundle-Activator: us.nineworlds.xstreamer.Activator
 Bundle-ActivationPolicy: lazy
