+++ com.example.rcpapp/src/us/nineworlds/xstreamer/Application.java
@@ -1,4 +1,4 @@
-package com.example.rcpapp;
+package us.nineworlds.xstreamer;
 
 
 import org.eclipse.equinox.app.IApplication;
