+++ com.example.rcpapp/src/us/nineworlds/xstreamer/ApplicationActionBarAdvisor.java
@@ -1,4 +1,4 @@
-package com.example.rcpapp;
+package us.nineworlds.xstreamer;
 
 import org.eclipse.jface.action.IMenuManager;
 import org.eclipse.jface.action.MenuManager;
