+++ com.example.rcpapp/src/us/nineworlds/xstreamer/ApplicationWorkbenchWindowAdvisor.java
@@ -1,4 +1,4 @@
-package com.example.rcpapp;
+package us.nineworlds.xstreamer;
 
 import org.eclipse.swt.graphics.Point;
 import org.eclipse.ui.application.ActionBarAdvisor;
